Creating a common URID crate
