+++ core/Cargo.toml
@@ -14,6 +14,7 @@ travis-ci = { repository = "RustAudio/rust-lv2", branch = "master" }
 maintenance = { status = "passively-maintained" }
 
 [dependencies]
+urid = "0.1.0"
 lv2-sys = "1.0.0"
 lv2-core-derive = "1.0.0"
