+++ core/src/prelude.rs
@@ -4,4 +4,3 @@ pub use crate::match_extensions;
 pub use crate::plugin::{lv2_descriptors, Plugin, PluginInfo, PortCollection};
 pub use crate::port::*;
 pub use crate::sys::LV2_Descriptor;
-pub use crate::{Uri, UriBound};
