+++ urid/Cargo.toml
@@ -16,7 +16,6 @@ maintenance = { status = "passively-maintained" }
 [dependencies]
 lv2-core = "1.0.0"
 lv2-sys = "1.0.0"
-lv2-urid-derive = "1.0.0"
 
 [features]
 default = []
