+++ urid/derive/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "lv2-urid-derive"
-version = "1.0.1"
+name = "urid-derive"
+version = "0.1.0"
 authors = ["Jan-Oliver 'Janonard' Opdenhövel <jan.opdenhoevel@protonmail.com>", "Adrien Prokopowicz <adrien.prokopowicz@gmail.com>"]
 edition = "2018"
 license = "MIT OR Apache-2.0"
