+++ urid/derive/src/urid_collection_derive.rs
@@ -18,8 +18,8 @@ pub fn urid_collection_derive_impl(input: TokenStream) -> TokenStream {
         .map(|ident| quote! {#ident: map.populate_collection()?,});
 
     let implementation = quote! {
-        impl ::lv2_urid::URIDCollection for #struct_name {
-            fn from_map(map: &::lv2_urid::Map) -> Option<Self> {
+        impl ::urid::URIDCollection for #struct_name {
+            fn from_map<M: ::urid::Map + ?Sized>(map: &M) -> Option<Self> {
                 Some(Self {
                     #(#field_inits)*
                 })
